requirements.txt
# xpu_benchmark dependencies
torch>=2.1.0
numpy>=1.24.0
triton>=3.0.0
cupti-python>=13.0.0
matplotlib>=3.7.0
